requirements.txt
Flask
Flask-MySQLdb
Flask-SQLAlchemy
flask-login
flask-mail
pydantic
pydantic[email]
python-dotenv